I3 affine point (#18)
* Affine point added

* Clippy

* Point and AffinePoint arithmetic by macro

* Formatting

* allow some buggy clippy warnings

* From used instead of into_*

* Stack overflow.

* Some minor fixes.

* Restore original state.

* Restored inline flags.
