+++ tom256/src/proofs/mod.rs
@@ -46,7 +46,7 @@ impl<C: Curve, CC: Cycle<C>> ZkAttestProof<C, CC> {
         let r_inv = input.signature.r.inverse();
         let u1 = s_inv * input.msg_hash;
         let u2 = s_inv * input.signature.r;
-        let r_point = Point::<C>::GENERATOR.double_mul(&u1, &input.pubkey, &u2);
+        let r_point = Point::<C>::GENERATOR.double_mul(&u1, &Point::from(&input.pubkey), &u2);
         let s1 = r_inv * input.signature.s;
         let z1 = r_inv * input.msg_hash;
         let q_point = &Point::<C>::GENERATOR * z1;
